refactor(google_vertex): extract auth/SDK-init slice to google_vertex_auth.py
Move the deferred auth group (lazy SDK/protobuf imports, OAuth2 access-token
cache, temp-credential prep, initialize_vertex_ai) out of google_vertex_client.py
into google_vertex_auth.py. Client drops 1101 -> 913 lines (-17%); combined with
the earlier catalog/transforms split, the original 1785-line client is now 913
(-49% overall).

- The mutable SDK/token globals (_vertexai, _TOKEN_CACHE, ...) are owned by the
  auth module and only touched by these functions; the request impls consume SDK
  objects / tokens via the functions' return values, so the move is behavior-neutral.
- The 5 auth functions + _DEFAULT_TRANSPORT are re-imported into
  google_vertex_client so src.services.google_vertex_client.<fn> stays patchable;
  the request impls call them by bare name -> resolve to the client binding ->
  existing test patches (multi_endpoint, tools) still bite. Verified: all 24
  auth-patching canaries pass.
- Removed now-orphaned client imports.

Verified against the full local suite (CI billing-locked): zero new regressions on
the google_vertex / chat / route surfaces vs the clean-HEAD baseline; only delta is
a known missing-API-key xai flake in an untouched file.
